Fixed validator stats and updated inactivity score
Removed faulty query for validator_balance in the validator stats table. Updated inactivity stat to show the current maximum value rather than the average value. With many validators the average could be very small, which may not be as meaningful as knowing that a single validator has a very bad score.